pyproject.toml
[project]
name = "hello-py"
version = "0.1.0"
description = "My Project API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
    
    # Data Validation & Serialization
    "pydantic>=2.0.0",
    
    # Environment Configuration
    "python-dotenv>=1.0.0",
    
    # Logging
    "loguru>=0.7.0",
    
    # Kafka
    "confluent-kafka>=2.3.0",
    
    # Celery & Task Queue
    "celery>=5.3.0",
    "redis>=5.0.0",
    "flower>=2.0.0",
    
    # Utilities
    "python-multipart>=0.0.6",  # For FastAPI file uploads
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    
    # Code Quality
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    
    # Development Tools
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["org", "boss", "celery_task", "kafka_poc"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["org", "boss", "celery_task", "kafka_poc"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]